Update draft-holmer-rmcat-transport-wide-cc-extensions-01.txt
I'm not sure anymore, that it's a good idea to force the receiver to re-send statuses should the window slide backwards. Probably it would be better to always set these messages as "no received" if the status for the same sequence number shall be sent again. This will simplify for the receiver a lot, as it will not have to keep a history of already sent statuses... And the receiver will not have to keep a history if it wants to detect duplicates. WDYT?